moved classes to separate package removed redundant empty lines
