+++ jjdz2-email/src/main/java/com/jbd/cutEmails/RudeWordsInContent.java
@@ -1,4 +1,5 @@
-package com.jbd;
+package com.jbd.cutEmails;
+import com.jbd.Email;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.Marker;
