+++ jjdz2-email/src/main/java/com/jbd/cutEmails/TwoMailsNoAnswer.java
@@ -1,5 +1,6 @@
-package com.jbd;
+package com.jbd.cutEmails;
 
+import com.jbd.Email;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.Marker;
@@ -21,33 +22,23 @@ public class TwoMailsNoAnswer{
     public LocalDateTime recieverMailDate;
     public LocalDateTime ourOneBeforeLastDate;
 
-    void addingAdressesToList(List<Email> eMailKeeper) {
-
+    public void addingAdressesToList(List<Email> eMailKeeper) {
         for (Email email : eMailKeeper) {
-
             adresses.add(email.getFrom());
         }
-
-
     }
 
-    void addingDatesToList(List<Email> eMailKeeper) {
-
+    public void addingDatesToList(List<Email> eMailKeeper) {
         for (Email email : eMailKeeper) {
-
-
             dates.add(email.getData());
         }
         Collections.sort(dates);
-
-
     }
 
     public void removingUserMailFromList(List<Email> eMailKeeper) {
         for (Email e : eMailKeeper) {
             adresses.removeIf(user -> user.contains(userMail));
         }
-
     }
 
     public void decideIfTwoAnswers() {// This code waits for else feature
