+++ jjdz2-email/src/main/java/com/jbd/cutEmails/WeirdMessagesServlet.java
@@ -1,5 +1,4 @@
-package com.jbd;
-
+package com.jbd.cutEmails;
 
 import javax.ejb.Stateless;
 
@@ -8,9 +7,6 @@ public class WeirdMessagesServlet{
 
     public String stringi(){
         String olo = "dupsko";
-
         return olo;
     }
-
-
 }
